calendar form available via ajax

+++ src/cal/views.py
@@ -29,34 +29,35 @@ def calendar_view(request, template_name='calendar/calendar_view.html'):
 
 
 @login_required
-def calendar_create(request, template_name='calendar/calendar_form.html'):
-	form = CalendarForm(request.POST or none)
-	if form.is_valid():
+def calendar_create(request):
+	form = CalendarForm(request.POST or None)
+	if form.is_valid() and request.method == "POST" and request.is_ajax:
 		calendar = form.save(commit=False)
 		calendar.calendar_owner = request.user
 		calendar.save()
 
-		return redirect('calendar_view', pk=calendar.id)
+		return HttpResponse("data saved")
 
-	return render(request, template_name, {'form':form})
+	return render(request, 'calendar/calendar_form.html', {'form':form})
+	
 
 
 @login_required
 def calendar_update(request, pk, template_name='calendar/calendar_form.html'):
-        calendar = get_object_or_404(Adress, pk=pk)
-        form = AdressForm(request.POST or None, instance=calendar)
-        if form.is_valid():
+        calendar = get_object_or_404(Calendar, pk=pk)
+        form = CalendarForm(request.POST or None, instance=calendar)
+        if form.is_valid() and request.method == "POST" and request.is_ajax:
                 form.save()
-                return redirect('calendar_list')
+                return HttpResponse("data saved")
 
         return render(request, template_name, {'form':form})
 
 @login_required
 def calendar_delete(request, pk, template_name='calendar/calendar_delete.html'):
         calendar = get_object_or_404(Calendar, pk=pk)
-        if request.method=='POST':
+        if request.method=='POST' and request.is_ajax:
                 calendar.delete()
-                return redirect('calendar_list')
+                return HttpResponse("data deleted")
 
         return render(request, template_name, {'object':calendar})
